# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Padel Match API",
        "description": "Documentação da API do Padel Match",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/players": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Lista todos os players",
                "description": "Retorna os dados de todos os players cadastrados, com filtros opcionais",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "parameters": [
                    {
                        "name": "full_name",
                        "in": "query",
                        "description": "Filtrar por nome do jogador (busca parcial)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "João"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filtrar por nível do jogador",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "side",
                        "in": "query",
                        "description": "Filtrar por lado do jogador",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "left",
                                "right",
                                "both"
                            ],
                            "example": "right"
                        }
                    },
                    {
                        "name": "uf",
                        "in": "query",
                        "description": "Filtrar por UF (sigla do estado)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "SP"
                        }
                    },
                    {
                        "name": "municipio_ibge",
                        "in": "query",
                        "description": "Filtrar por código IBGE do município (7 dígitos)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "3550308"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de players",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/player/id": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Exibe o player",
                "description": "Retorna os dados do player solicitado pelo id",
                "operationId": "cb4ef59d666f6630d465c848639befe1",
                "responses": {
                    "200": {
                        "description": "Player encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/player": {
            "put": {
                "tags": [
                    "Player"
                ],
                "summary": "Atualiza o player do usuário",
                "description": "Atualiza os dados do player associado ao usuário autenticado",
                "operationId": "2e20018208411a2682bf9a1b18babba8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "level",
                                    "side"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "(49) 99999-9999"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "side": {
                                        "type": "string",
                                        "enum": [
                                            "left",
                                            "right",
                                            "both"
                                        ],
                                        "example": "both"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Jogador intermediário"
                                    },
                                    "profile_image_url": {
                                        "type": "string",
                                        "example": "https://site.com/foto.jpg"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "municipio_ibge": {
                                        "type": "string",
                                        "example": "3550308"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Player atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Player não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Cadastra um novo player",
                "description": "Cria o perfil de player para o usuário autenticado",
                "operationId": "287a67768b206bd753d4b762bc2876a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "level",
                                    "side"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "João da Silva"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "(49) 99999-9999"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "side": {
                                        "type": "string",
                                        "enum": [
                                            "left",
                                            "right",
                                            "both"
                                        ],
                                        "example": "right"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Jogador iniciante"
                                    },
                                    "profile_image_url": {
                                        "type": "string",
                                        "example": "https://site.com/foto.jpg"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "municipio_ibge": {
                                        "type": "string",
                                        "example": "3550308"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Player criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jogador não encontrado"
                    },
                    "409": {
                        "description": "Usuário já possui um player cadastrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me/player": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Player vinculado ao usuário",
                "description": "Retorna se tem player vinculado ao usuário autenticado",
                "operationId": "724b279740dca85168d3716dc038e628",
                "responses": {
                    "200": {
                        "description": "Dados do player",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não possui player vinculado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}